+++ src/tests/SimplexTest.java
@@ -16,10 +16,10 @@ public class SimplexTest {
 	@Test
 	public void results_1() throws Exception {
 		int objective = Expression.OBJECTIVE_MAX;
-		int[] objectiveFunction = {80, 60};
-		int[][] constraints = {{4, 6}, {4, 2}, {0, 1}};
-		int[] constraintSigns = {Expression.CONSTRAINT_SIGN_GT, Expression.CONSTRAINT_SIGN_LT, Expression.CONSTRAINT_SIGN_LT};
-		int[] b = {24, 16, 3};
+		double[] objectiveFunction = {80, 60};
+		double[][] constraints = {{4, 6}, {4, 2}, {0, 1}};
+		double[] constraintSigns = {Expression.CONSTRAINT_SIGN_GT, Expression.CONSTRAINT_SIGN_LT, Expression.CONSTRAINT_SIGN_LT};
+		double[] b = {24, 16, 3};
 		
 		Expression e = new Expression(objective, objectiveFunction, constraints, constraintSigns, b);
 		Simplex s = Simplex.getInstance();
@@ -31,10 +31,10 @@ public class SimplexTest {
 	@Test
 	public void results_2() throws Exception {
 		int objective = Expression.OBJECTIVE_MIN;
-		int[] objectiveFunction = {80, 60};
-		int[][] constraints = {{4, 6}, {4, 2}, {0, 1}};
-		int[] constraintSigns = {Expression.CONSTRAINT_SIGN_GT, Expression.CONSTRAINT_SIGN_LT, Expression.CONSTRAINT_SIGN_LT};
-		int[] b = {24, 16, 3};
+		double[] objectiveFunction = {80, 60};
+		double[][] constraints = {{4, 6}, {4, 2}, {0, 1}};
+		double[] constraintSigns = {Expression.CONSTRAINT_SIGN_GT, Expression.CONSTRAINT_SIGN_LT, Expression.CONSTRAINT_SIGN_LT};
+		double[] b = {24, 16, 3};
 		
 		Expression e = new Expression(objective, objectiveFunction, constraints, constraintSigns, b);
 		Simplex s = Simplex.getInstance();
@@ -49,16 +49,16 @@ public class SimplexTest {
 		exception.expectMessage("Unlimited solution.");
 		
 		int objective = Expression.OBJECTIVE_MIN;
-		int[] objectiveFunction = {20, 16, 24, 10, 10, 8, 12, 18, 10};
-		int[][] constraints = {	{-1, -1, -1, 0, 0, 0, 0, 0, 0},
+		double[] objectiveFunction = {20, 16, 24, 10, 10, 8, 12, 18, 10};
+		double[][] constraints = {	{-1, -1, -1, 0, 0, 0, 0, 0, 0},
 								{0, 0, 0, -1, -1, -1, 0, 0, 0},
 								{0, 0, 0, 0, 0, 0, -1, -1, -1},
 								{1, 0, 0, 1, 0, 0, 1, 0, 0},
 								{0, 1, 0, 0, 1, 0, 0, 1, 0},
 								{0, 0, 1, 0, 0, 1, 0, 0, 1}
 							  };
-		int[] constraintSigns = {Expression.CONSTRAINT_SIGN_EQ, Expression.CONSTRAINT_SIGN_EQ, Expression.CONSTRAINT_SIGN_EQ, Expression.CONSTRAINT_SIGN_EQ, Expression.CONSTRAINT_SIGN_EQ, Expression.CONSTRAINT_SIGN_EQ};
-		int[] b = {-300, -500, -200, 300, 400, 300};
+		double[] constraintSigns = {Expression.CONSTRAINT_SIGN_EQ, Expression.CONSTRAINT_SIGN_EQ, Expression.CONSTRAINT_SIGN_EQ, Expression.CONSTRAINT_SIGN_EQ, Expression.CONSTRAINT_SIGN_EQ, Expression.CONSTRAINT_SIGN_EQ};
+		double[] b = {-300, -500, -200, 300, 400, 300};
 		
 		Expression e = new Expression(objective, objectiveFunction, constraints, constraintSigns, b);
 		Simplex.getInstance().start(e);
@@ -70,8 +70,8 @@ public class SimplexTest {
 		exception.expectMessage("Permissible solution does not exist.");
 		
 		int objective = Expression.OBJECTIVE_MIN;
-		int[] objectiveFunction = {4, 6, 7, 3, 4, 5, 6, 7, 8, 8, 5, 3};
-		int[][] constraints = {	{-1, -1, -1, 0, 0, 0, 0, 0, 0, 0, 0, 0},
+		double[] objectiveFunction = {4, 6, 7, 3, 4, 5, 6, 7, 8, 8, 5, 3};
+		double[][] constraints = {	{-1, -1, -1, 0, 0, 0, 0, 0, 0, 0, 0, 0},
 								{0, 0, 0, -1, -1, -1, 0, 0, 0, 0, 0, 0},
 								{0, 0, 0, 0, 0, 0, -1, -1, -1, 0, 0, 0},
 								{1, 0, 0, 1, 0, 0, 1, 0, 0, -1, 0, 0},
@@ -79,8 +79,8 @@ public class SimplexTest {
 								{0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, -1},
 								{0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1}
 							  };
-		int[] constraintSigns = {Expression.CONSTRAINT_SIGN_EQ, Expression.CONSTRAINT_SIGN_EQ, Expression.CONSTRAINT_SIGN_EQ, Expression.CONSTRAINT_SIGN_EQ, Expression.CONSTRAINT_SIGN_EQ, Expression.CONSTRAINT_SIGN_EQ, Expression.CONSTRAINT_SIGN_EQ};
-		int[] b = {-90, -20, -10, 30, 15, 15, 50};
+		double[] constraintSigns = {Expression.CONSTRAINT_SIGN_EQ, Expression.CONSTRAINT_SIGN_EQ, Expression.CONSTRAINT_SIGN_EQ, Expression.CONSTRAINT_SIGN_EQ, Expression.CONSTRAINT_SIGN_EQ, Expression.CONSTRAINT_SIGN_EQ, Expression.CONSTRAINT_SIGN_EQ};
+		double[] b = {-90, -20, -10, 30, 15, 15, 50};
 		
 		Expression e = new Expression(objective, objectiveFunction, constraints, constraintSigns, b);
 		Simplex s = Simplex.getInstance();
